requirements.txt
alembic==1.8.1
anyio==3.6.1
Authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyshorteners==1.0.1
python-dotenv==0.20.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
zipp==3.8.1
